feat: 精简代码
