--- core/plugin/config.go
@@ -1,6 +1,0 @@
-package plugin
-
-type Config struct {
-	Name   string            `yaml:"name,omitempty"`
-	Params map[string]string `yaml:"params,omitempty"`
-}
