--- core/plugin/registry.go
@@ -1,6 +1,0 @@
-package plugin
-
-type Registry interface {
-	RegisterBean(name string, bean interface{}) error
-	GetBean(name string) interface{}
-}
