--- core/server/config.go
@@ -1,6 +1,0 @@
-package server
-
-type Config struct {
-	Protocol string `yaml:"protocol,omitempty"`
-	Addr     string `yaml:"addr,omitempty"`
-}
