--- core/server/constant.go
@@ -1,6 +1,0 @@
-package server
-
-const (
-	PluginName = "server"
-	ManagerName = "server.manager"
-)
