--- core/shorturl/constant.go
@@ -1,6 +1,0 @@
-package shorturl
-
-const (
-	PluginName = "idgens"
-	ShortUrlManager = "idgens"
-)
