+++ core/shorturl/setup.go
@@ -9,7 +9,13 @@ import (
 
 func setup(ctx *plugin.Context) error {
 	shorturl := &service.Manager{}
-	ctx.RegisterBean(ShortUrlManager, shorturl)
+	ctx.RegisterBean(ManagerName, shorturl)
+
+	// shorturl extpoint plugin
+	su := ctx.GetParams().(*Config)
+	ctx.ApplyPlugin(su.IDGen.Name, su.IDGen.Params)
+	ctx.ApplyPlugin(su.IDEncode.Name, su.IDEncode.Params)
+	ctx.ApplyPlugin(su.IDStore.Name, su.IDStore.Params)
 
 	mgr := server.GetManager(ctx)
